refactor(mesh-v2): extract event-sender mixin
issue #566 の責務分割 8/N。Scratch broadcast を mesh 経由で送信するロジック
を `event-sender.js` に切り出し。

切り出し対象:
- fireEvent (キューイング、重複検出、サイズ制限の FIFO 削除)
- _generateOrderKey (issue #556 の 7 桁連番付き orderKey 生成)
- startEventBatchTimer / stopEventBatchTimer
- processBatchEvents (1000 件単位の分割送信)
- fireEventsBatch (WebSocket: FIRE_EVENTS / Polling: RECORD_EVENTS)
- reportEventStatsIfNeeded (キュー統計の定期出力)

mesh-service.js: 685 → 504 行 (-181)
event-sender.js: 207 行 (新規)
